create separated mod for reparsing functionality
